JSONViewer.js and JSONViewer.css are moved to JSONViewer folder
