+++ practice-app/frontend/practice-app/src/Containers/MainContainer/MainContainer.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
 import "./MainContainer.scss";
-import JSONViewer from "./JSONViewer"
+import JSONViewer from "../JSONViewer/JSONViewer.js"
 
 function MainContainer() {
     return <div className="main-container">
